requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncio-pool==0.6.0
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.42.0
blis==1.3.0
Brotli==1.1.0
cachetools==6.1.0
catalogue==2.0.10
cbor==1.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cobble==0.1.4
confection==0.1.5
coverage==7.10.3
cryptography==45.0.5
cymem==2.0.11
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.9
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
FlagEmbedding==1.3.5
frozenlist==1.7.0
fsspec==2025.3.0
geomet==1.1.0
gradio==5.43.1
gradio_client==1.12.1
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.7
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
ijson==3.4.0
iniconfig==2.1.0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
llama_cpp_python==0.3.15
logger==1.4
lxml==5.3.0
lz4==4.4.4
Mako==1.3.10
mammoth==1.10.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mlx==0.28.0
mlx-lm==0.26.3
mlx-metal==0.28.0
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openai==1.97.1
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
pandas==2.2.3
pandas-stubs==2.3.0.250703
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.0
peft==0.17.1
pgvector==0.3.6
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readability==0.3.1
redis==6.2.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rsa==4.9.1
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
scylla-driver==3.29.3
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
tenacity==9.0.0
textstat==0.7.4
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
trec-car-tools==2.6
typer==0.16.1
types-pytz==2025.2.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.35.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
